Add unique states bitmask helpers to `rivershared/uniquestates`
This one's aimed at a bug discovered in Pro wherein because the project
didn't have access to the unique states bitmask helpers, the
`UniqueStates` property was left off job rows. Here, put the headers in
a general location in `rivershared` that can be accessed from both
projects.
